feat: offchain cancel and tokenset bid tests
